requirements.txt
# PRD targets Python 3.11 + CUDA 12.1; use flexible pins so 3.12+ / Windows wheels resolve.
# For legacy MediaPipe ``mp.solutions.pose``, prefer Python 3.11 and mediapipe<0.10.14.
# Python 3.12+ uses MediaPipe Tasks (pose landmarker .task) — see configs/pipeline.yaml.

# Core inference
ultralytics>=8.2.0
mediapipe>=0.10.30
onnxruntime-gpu>=1.18.0

# Data and annotation
pycocotools>=2.0.7
label-studio>=1.12.0
label-studio-sdk
dvc[s3]>=3.50.0
roboflow

# ML stack
torch>=2.3.0
torchvision
xgboost>=2.0.3
shap>=0.45.0
mlflow>=2.13.0
optuna>=3.6.0
scikit-learn

# Utilities
opencv-python-headless
numpy
scipy
pandas
tqdm
rich
pyyaml
ffmpeg-python

# Serving
fastapi
uvicorn

# Testing
pytest
